requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
faiss-cpu
langchain-ollama
